Adjust abstract models structure in `helpers`
